requirements.txt
streamlit
Pillow
# aggiungi qui altre librerie se le usi (es. numpy, rembg, opencv-python)